----------------------------------------------------------------------  Modified Files:  	Epetra_CrsGraph.h Epetra_CrsMatrix.h  	Epetra_LinearProblemRedistor.cpp  	Epetra_LinearProblemRedistor.h  ----------------------------------------------------------------------
Partially addresses bug 57.

This mod contains the first legitimate version of Epetra_LinearProblemRedistor,
a class for constructing linear problem objects that can be redistributions
of existing problems, including replication on all or some processors on a
parallel machine, on-the-fly transposition of the matrix, and extraction of
the matrix, and vector data into raw arrays for use with solvers that expect
a Harwell-Boeing style of data structures.

Changes to CrsGraph were made to allow cheap inline operator[] notation
similar to what has been in CrsMatrix for some time.  Changes to CrsMatrix
just fixed a few comments.