Added a few things to Makefile.(platform).  At this time only testSuite/unit/Makefile uses the new macros.  You'll eventually need to copy some of the new stuff into your Makefile.customize.
First, I added TSTT_LINK, which is the portion of the link
line that links in a TSTT implementation.  Second, I added
AOMD_TSTT_LINK and MDB_TSTT_LINK, which are the link lines
for AOMD and MDB respectively.  If you want to change which
implementation you are linking in, change the TSTT_LINK
accordingly.

Second, I added MSQ_BASE_DIR, which allows Makefile.customize
to be included in makefiles located somewhere other than the
root mesquite directory.