Fix to Bug# 138.
Now Doxygen will work even if the build is done in another
directory. The documentation is still created in the original source
directory as opposed to the build directory. The file created by
configure is copied over tot he orignial source location before we run
Doxygen.

I've also added a README file which explains how to generate the
documentation, even without running configure (since it's the
documentation that explains how to run configure and configure that
generates the documentation, we were having a catch-22 situation).