Re-vamped the sublists as we discussed in the last developer's meeting.
The new structure of the sublists is probably most clearly seen in the
NOX Parameter Reference Page generated by Doxygen.

I have fixed the Linear and Rosenbrock LAPACK examples. All the other
examples still need to be fixed. They will compile and execute, but
some of the options may be ignored if they aren't in the right
place. The best way to see if your options are unused is to execute

solver.getParameterList().print(cout);

Roger has told me he will fix the Epetra examples over the weekend.

Russ is on vacation, but hopefully he can fix the Petsc examples when
he returns.

Andy and Eric, hopefully this is the last of the major changes that
will cause major pain and confusion for LOCA. Sorry 'bout this!!