overloaded the get_next_patch function. We now have get_next_patch( ..., PachDataUser*, ...) that delegates to get_next_patch(..., PatchDataParameters&, ...) where the code is implemented.
I hadn't bothered doing that before, but since we need it in unit tests where Patches are created separately from the PatchDataUser, I decided to give the option to use both signatures. This avoid the dirty trick "create a dummy PatchDataUser" and allows to keep the constructor of the base class (PatchDataUser) protected, as it should be.