Changed GENERATE_TAGFILE in Doxyfile.in, Doxyfile.default to
GENERATE_TAGFILE       = nox.tag

to tell doxygen to generate a tag file, nox.tag, when the documentation is
generated.  This tag file is placed in the doc directory and allows nox
documentation to be externally linked into other packages without having
to regenerate the nox documentation.  This was done to separate the nox
documentation from the loca documentation.