Added PatchData::snap_vertex_to_domain(size_t vertex_index,                                        MsqError &err);
The function doesn't do anything right now, but at least you
can start using it in the algorithms.  The more I talk to
people, the more unsure I am of the right way to do geometry
operations.  The only thing I'm sure of is that the snap()
function in the TSTT mesh modification interface is the wrong
way to do it...