----------------------------------------------------------------------
 Modified Files:
 	src/Ifpack_IlukGraph.cpp
 ----------------------------------------------------------------------

Partially addresses bug 214.

Made a small code change to the IlukGraph class to explicitly define
the column map for an overlap graph object.  This ensures that the maps
of the rows and columns of the overlap graph are identical and improves
the factorization properties for overlapped matrices.

The Epetra_VbrMatrix does not work with this approach at this time. More
work needs to be done on the Import/Export filtering facilities of
the Epetra_VbrMatrix class.