Adding functionality to PatchData and MeshSet to utilize SimplifiedGeometryEngine (see commit message below).  A SimplifiedGeometryEngine pointer can now be stored within MeshSet, and this pointer is passed to PatchData whenever the get_next_patch is called.  Adding data members and member functions to access the data.  Along with the pointer, MeshSet and PatchData also store a GeometryEngine enumeration.
Adding a class to store some simple geometry information.  This
class, SimplifiedGeometryEngine, will have the ability to return
surface normal and project points to the surface for PLANEs and
for SPHEREs.  The user can create this object and associate it
with a MeshSet.  For the mesh quads_on_sphere_529.vtk which is
a mesh of sphere (radius three and centered at (2,2,0)), the
following code can be used to create and associate the
SimplifiedGeometryEngine object:

/**************** CREATE GEOMETRY **************************/
    //create vector3d signifying the center of the radius
  Vector3D center(2,2,0);
    //double for the radius of the sphere
  double sphere_rad=3.0;
    //the Geometry object
  SimplifiedGeometryEngine msq_geom;
    //set the geometry to the appropriate sphere
  msq_geom.set_geometry_to_sphere(center,sphere_rad,err);
  MSQ_CHKERR(err);
    //add the geomtry to the MeshSet
  mesh_set1.set_simplified_geometry_engine(&msq_geom);
/*************** END CREATE GEOMETRY ***********************/

This code has not been tested thoroughly yet.  However, the
sphere case has been tested on a few examples.  The planar
case currently uses a temporary solution for
snap_vertex_on_domain and will not be robust until this is
replaced with a more permanent solution.