Added "volatile" keyword in Box_Assign3 and Box_Assign2 to remove strange behavior for RIB Box Assign on linux using gcc -O. (Bob's suggestion)
ch_film, zdrive.inp.rib-partless, missed a partition in Box_Assign
due to roundoff (or something) in Box_Assign2.  Point_Assign would
return the missed partition for a corner of the box used in Box_Assign.
This situation is an error; processors and partitions returned by
Point_Assign should also be returned by Box_Assign if the point is within
the box.