----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	packages/aztecoo/aclocal.m4 packages/aztecoo/configure
 	packages/aztecoo/configure.ac
 	packages/aztecoo/example/Makefile.am
 	packages/aztecoo/example/Makefile.in
 	packages/aztecoo/example/AztecOO/Makefile.am
 	packages/aztecoo/src/AztecOO_config.h.in
 	packages/aztecoo/src/az_aztec_defs.h
 	packages/aztecoo/src/az_check.c
 	packages/aztecoo/src/az_solve.c
 	packages/aztecoo/src/az_subdomain_solver.c
 	packages/triutils/aclocal.m4 packages/triutils/configure
 	packages/triutils/configure.ac
 	packages/triutils/src/Triutils_config.h.in
 Added Files:
 	packages/aztecoo/example/AztecOO/Makefile.in
 ----------------------------------------------------------------------
Addresses bug 106.

When using autotools, y12m is no longer needed by default.  This means
that the AZ_lu preconditioner is no longer available by default.  To
turn the preconditioner back on, add --with-aztecoo-azlu to the configure
line.  For now, --enable-referencey12m is still available to create the
y12m library for those who need it.

Please note that this change has not, and is not planned to be made in
Trilinos 3.0.


The 'AztecOO' aztecoo example is now autotooled.  The example was used
to test the fix for bug 106.

Finally, the check for TFLOPS was removed from the configure.ac files for
AztecOO and Triutils.  This check does not yet work properly, and is not
needed in either of these packages.