A couple of additions to the StopWatch class.  Adding a counter which is incremented each time the StopWatch is started.  For StopWatches that are used for timing functions, this can be used to get the number of times the function was called.  Also adding member functions to access this data.
Adding a function StopWatchCollection::get_keys_sorted_by_time
(std::vector<StopWatchCollection::Key> sorted_keys).  This
function fills a vector of Keys.  The key associated with
the StopWatch which ran the longest time is the first
slot of the vector, and the key associated with the StopWatch
that ran the shortest amount of time is in the last vector
slot.  This is mainly for printing purposes.

Adding a function PRINT_TIMING_DIAGNOSTICS() which is currently
in MsqMessage.  I am not sure where this function actually
belongs; so I placed it there for the time being.  Feel free
to move it.  For each StopWatch in the global collection, the
function prints the total time, number of times started, and
the timer's name (it doesn't print the percentage of total
time, yet).