----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	aclocal.m4 configure configure.ac config/tac_arg_check_mpi.m4
 	config/tac_arg_config_mpi.m4 config/tac_arg_with_libdirs.m4
 ----------------------------------------------------------------------
Fixes bugs 236, 237, and 239.

Some configure line options were not the same for Nox and the rest of
Trilinos.  In particular, mpi options were conflicting, and two options
served the same purpose for the majority of Trilinos, while only one
(the one used less frequently) was valid for Nox.

Here is a summary of the changes:

--with-mpi-include=-I<DIR> has been removed.  Use --with-mpi-incdir=DIR
instead.  --with-mpi-include was originally taken from Nox, but Nox was
updated and the rest of Trilinos was not.  NOTE: --with-mpi-incdir does not
accept a -I at the beginning of the path as --with-mpi-include did.  ie use
--with-mpi-incdir=/usr/path/include

--with-mpi-libdir no longer accepts -L at the beginning of the path.
This was a conflict with Nox and was causing problems because the syntax
for the two options was not compatible.  ex. --with-mpi-libdir=/usr/path/lib

--with-libdirs should no longer be used.  Nox does not accept this option.
Use --with-ldflags instead.  ex. --with-ldflags="-L/usr/dir -L/usr/dir2"


A number of help strings were changed to account for this change.  Also,
some other help strings that have proved to be misleading or difficult for
users to understand have been reworded.  These improvements address the
most common areas that users have had trouble with when configuring Trilinos.
Examples were added for some of these trouble spots.

Until now, --disable-mpi has had the same effect as --enable-mpi, thank you to
Ken S. for pointing this out.  This problem is resoved with this update.

New package also can be configured with the rest of Trilinos, for now
the default is not to configure and build.  However, there is talk of
building it by default in the future.  It serves as a good sanity check,
and the library is VERY small.

Note that these changes have only been applied to the development branch.

Test harness passed.