----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	packages/epetra/src/Epetra_LAPACK.cpp
 	packages/epetra/src/Epetra_LAPACK.h
 ----------------------------------------------------------------------
Fixes bug 242.

Rewrote the wrappers for the dlamch and slamch Lapack routines.  The
wrappers did not work for Intel compilers using the CXML libraries.
The old wrappers took in a char and returned a double (DLAMCH) or a
float (SLAMCH).  The wrappers did not work for Intel CXML because the
name of the Epetra::LAPACK method was the same Lapack routine using a
simple upercase name mangling scheme.  The new wrappers, both called
LAMCH, are void methods and take in the char, and either a double
(double precision) or a float (single precision).  The value that was
formerly returned is now passed back in the second argument.
This is more consistant with the rest of the Blas and Lapack wrappers.

All tests pass on Linux and Intel CXML.